fix: Исправлена клавиатура категорий

+++ app/bot/handlers/client/documents.py
@@ -2,17 +2,19 @@ from aiogram import types
 from aiogram.dispatcher import FSMContext
 
 from app.bot.handlers.pagination_handler import pagination_objects
-from app.bot.keyboard import (create_category_keyboard,
+from app.bot.keyboard import (create_cancel_keyboard, create_category_keyboard,
                               create_registration_keyboard)
 from app.bot.states import ClientMenu
 from app.core.config import bot
-from app.core.db.repository import document_service, user_service
+from app.core.db.repository import (category_service, document_service,
+                                    user_service)
 
 
 async def files_command(message: types.Message, state: FSMContext):
     user = await user_service.get_user_by_telegram_id(
         telegram_id=message.from_user.id
     )
+    categories = await category_service.get_all_objects()
     if user:
         if user.is_banned:
             bot.send_message(
@@ -20,14 +22,23 @@ async def files_command(message: types.Message, state: FSMContext):
                 "<i>Доступ запрещен.</i>",
                 parse_mode=types.ParseMode.HTML,
             )
-        await bot.send_message(
-            message.from_user.id,
-            "<i>Выберите какие файлы вам нужны.</i>",
-            reply_markup=await create_category_keyboard(),
-            parse_mode=types.ParseMode.HTML,
-        )
-        await state.set_state(ClientMenu.select_documents.state)
-        await message.delete()
+        if len(categories) != 0:
+            await bot.send_message(
+                message.from_user.id,
+                "<i>Выберите какие файлы вам нужны.</i>",
+                reply_markup=await create_category_keyboard(categories),
+                parse_mode=types.ParseMode.HTML,
+            )
+            await state.set_state(ClientMenu.select_documents.state)
+            await message.delete()
+        else:
+            await bot.send_message(
+                message.from_user.id,
+                "<i>В данный момент файлов нет.</i>",
+                reply_markup=create_cancel_keyboard(),
+                parse_mode=types.ParseMode.HTML,
+            )
+
     else:
         await bot.send_message(
             message.from_user.id,
